Implement the close and commit methods.
This is in preparation for the handling of memory resident UCF documents. The
close method had to be implemented too so that it could be set to call the new
local version of the commit method instead of the one in ZipFile.

In actual fact the close method is an alias for the commit method as in the
underlying zip implementation the close method simply calls the commit method
with no additional logic.